\name{brada-package}
\alias{brada-package}
\alias{brada-package}
\docType{package}
\title{
\packageTitle{brada}
}
\description{
\packageDescription{brada}
}
\details{
Provides access to a range of functions for analyzing, applying and visualizing \cr
Bayesian response-adaptive trial designs for a binary endpoint. Includes the \cr
predictive probability approach and the predictive evidence value designs for \cr
binary endpoints.
\packageDESCRIPTION{brada}
\packageIndices{brada}

}
\author{
\packageAuthor{brada}

Maintainer: \packageMaintainer{brada}
}

\keyword{ package }
