\name{names.brada}
\alias{names.brada}
\alias{names}
\title{
names.brada
}
\description{
Plots the names of the objects stored in the \code{brada} object of a Bayesian response-adaptive design analysis.
}
\usage{
\method{names}{brada}(x)
}
\arguments{
  \item{x}{
    An Object of class \code{"brada"}.
  }
}
\details{
Plots the names of the objects stored in the \code{trials} object of a Bayesian response-adaptive design analysis.
}
\value{
  Returns a list of names.
}
\author{
  Riko Kelter
}