\name{show.brada}
\alias{show.brada}
\alias{show}
\title{
show.brada
}
\description{
Prints the main results of a Bayesian response-adaptive design analysis to the console.
}
\usage{
\method{show}{brada}(object)
}
\arguments{
  \item{object}{
    An Object of class \code{"brada"}.
  }
}
\details{
Shows the main results of a Bayesian response-adaptive design analysis stored in an object of class \code{brada}.
}
\value{
  Prints the results onto the console.
}
\author{
  Riko Kelter
}